# docs/owner/phase9_5a/openapi.yaml
openapi: 3.0.3
info:
  title: Aura Owner Operations API
  version: 1.0.0
  description: 'Phase 9.5A foundation contract for the commercial-operations / employee-portal / admin-portal
    surface, extended by Phase 9.5D (Commercial Sales) and Phase 9.5E (Expenses/Payees/Cash-Closing/
    Report-Snapshots/Management-Notes). Distinct from /api/licensing/v1 (external product licensing,
    unchanged). x-status distinguishes planned (foundation-only, not routed) from implemented (real route
    in owner/app/__init__.py, documented from actual behavior -- see each path''s own docstring reference).
    Mobile auth (/auth/mobile/*) remains planned; mobile-ready schema, mobile app not built.

    '
servers:
- url: /api/operations/v1
security:
- bearerAuth: []
- cookieAuth: []
paths:
...
  /customers/{customer_id}:
    parameters:
    - name: customer_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    get:
      x-status: planned
      summary: Customer detail -- see customer-detail-api-view-model.md
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDetail'
          description: Successful response
        '404':
          $ref: '#/components/responses/Error'
          description: Error response
  /customers/{customer_id}/locations:
    parameters:
    - name: customer_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    post:
      x-status: planned
      summary: Capture a customer location (customers.capture_location) -- explicit action only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationCapture'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerLocation'
          description: Successful response
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque access token (mobile) -- see mobile-session-contract.md
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...
  responses:
    Error:
      description: Standard error envelope -- see api-error-catalog.md
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
      - error_code
      - message
      - correlation_id
      properties:
        error_code:
          type: string
          enum:
          - AUTHENTICATION_REQUIRED
          - MFA_REQUIRED
          - RECENT_AUTH_REQUIRED
          - PERMISSION_DENIED
          - RECORD_NOT_ASSIGNED
          - RECORD_NOT_FOUND
          - VERSION_CONFLICT
          - IDEMPOTENCY_CONFLICT
          - VALIDATION_ERROR
          - INVALID_STATE_TRANSITION
          - PRICE_VERSION_INACTIVE
          - DEVICE_POLICY_VIOLATION
          - DUPLICATE_LEAD
          - DUPLICATE_CUSTOMER
          - COMMISSION_NOT_ELIGIBLE
          - REFRESH_TOKEN_INVALID
          - REFRESH_TOKEN_REUSE_DETECTED
          - INTERNAL_ERROR
        message:
          type: string
        correlation_id:
          type: string
        details:
          type: object
          nullable: true
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        legal_name:
          type: string
        lifecycle_status:
          type: string
        converted_from_lead_id:
          type: string
          format: uuid
          nullable: true
    CustomerDetail:
      allOf:
      - $ref: '#/components/schemas/Customer'
      - type: object
        properties:
          commercial:
            type: object
            nullable: true
          licensing:
            type: object
            nullable: true
          version:
            type: integer
    LocationCapture:
      type: object
      required:
      - source
      properties:
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        accuracy_meters:
          type: number
          nullable: true
        source:
          type: string
          enum:
          - GPS
          - NETWORK
          - MANUAL
          - IMPORTED
        manual_address:
          type: string
          nullable: true
    CustomerLocation:
      allOf:
      - $ref: '#/components/schemas/LocationCapture'
      - type: object
        properties:
          id:
            type: string
            format: uuid
          captured_at:
            type: string
            format: date-time
          verified:
            type: boolean
...